QOrganizer: Fix handling of item IDs
This commit heavily reworks the way that item IDs are passed around.
First we add a couple of static functions to QOrganizerEdsEngine, to
translate from EDS coordinates, which consiste of a source ID and an
item ID, to QtOrganizer coordinates, which consist of a single
QOrganizerItemId (inside which we encode both athe EDS source and item
ID).

Then, we stop using a QString to store the collection ID, but either use
the proper QOrganizerCollectionId class or, wherever possible, just the
EDS source ID (which corresponds to QOrganizerCollectionId::localId()).

Fixes https://github.com/ubports/ubuntu-touch/issues/869