QtOrganizer: fix a race condition on collection creation
This was reported as an EDS bug here:
https://gitlab.gnome.org/GNOME/evolution-data-server/issues/53

The problem is taht the old code was assuming (and maybe this was
working fine with older EDS versions?) that the "source-added" signal
from the source registry would be emitted before the
e_source_registry_create_sources() method returned. But this is not
always the case.